requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
flake8
black
pytest
python-dotenv>=0.5.1
ipkykernel

# data
streamlit
pandas
kedro-datasets

# graph
plotly
nbformat
